fix: move file and create notifier folder
