+++ lib/screen/weather_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_training/component/error_dialog.dart';
 import 'package:flutter_training/component/temperature_button.dart';
 import 'package:flutter_training/component/weather_forecast_panel.dart';
+import 'package:flutter_training/notifier/weather_screen_notifier.dart';
 import 'package:flutter_training/state/weather_screen_error_state.dart';
-import 'package:flutter_training/state/weather_screen_notifier.dart';
 
 class WeatherScreen extends ConsumerWidget {
   const WeatherScreen({super.key});
