+++ lib/state/weather_forecast_panel_state.freezed.dart
