+++ lib/state/weather_repository_ui_state.freezed.dart
