+++ lib/state/weather_request.dart
