+++ lib/state/weather_request.freezed.dart
